requirements.txt
streamlit
streamlit-mic-recorder
openai-whisper